Use double price and remove using namespace std
Change ActualizarPrecio parameter from int to double in producto.h so it matches the product price type and supports fractional prices. Also remove the "using namespace std;" from empresa.h to avoid polluting the global namespace from a header.

+++ producto.h
@@ -21,7 +21,7 @@ public:
 	int ObtenerCantidadVentas() const {return m_cantidadVendida;};
 	double ObtenerVentas() const {return m_ventas;};
 	void ActualizarStock(int cantidad) {m_stock = cantidad;};
-	void ActualizarPrecio(int cantidad) {m_precio = cantidad;};
+	void ActualizarPrecio(double cantidad) {m_precio = cantidad;};
 	void ActualizarVentas(int cantidad) {m_cantidadVendida = cantidad;};
 	void VenderProducto(int cantidad);
 	bool operator<(const Producto& otro) const;
